Tornando o cleaner mais robusto

+++ backend/src/tests/utils/cleaner.py
@@ -1,4 +1,5 @@
 from main import app
+from sqlalchemy import delete
 from starlette.testclient import TestClient
 
 from .aux_functions import get_csrf_token
@@ -16,14 +17,13 @@ def clean_database() -> None:
         for email in db_session.query(UserSchema.email).filter(UserSchema.email != "motorizen@efscode.com.br").all()
     ]
 
-    db_session.close()
-
     if not user_emails:
         print("Database already clean.")
         return
 
     client = TestClient(app)
     for email in user_emails:
+        print(f"Deleting user {email}...")
         csrf_token = get_csrf_token(client)
         response = client.post(
             "/token",
@@ -36,11 +36,18 @@ def clean_database() -> None:
         )
 
         token_data = response.json()
+        print(token_data)
 
-        client.delete(
-            "/users/delete-user",
-            headers={"Authorization": f"Bearer {token_data['access_token']}"},
-        )
+        if response.status_code == 200:
+
+            client.delete(
+                "/users/delete-user",
+                headers={"Authorization": f"Bearer {token_data['access_token']}"},
+            )
+
+        db_session.execute(delete(UserSchema).where(UserSchema.email == email))
+    db_session.commit()
+    db_session.close()
 
     print("Database cleaned.")
 
@@ -48,6 +55,10 @@ def clean_database() -> None:
 if __name__ == "__main__":
     import sys
 
-    clean_database()
+    try:
+        clean_database()
+    except Exception as e:
+        print(e)
+        sys.exit(1)
 
     sys.exit(0)
